Check in FiNesse pages with LF line endings, not CRLF
The new FitNesse wiki format in v20161106 does not recognize page
properties where the line ending is CRLF.  It only accepts LF as the
line ending.  See FitNesse issues #999, #1050, #1058.  This should be
fixed in an upcoming FitNesse release but until then the line endings
will have to be handled in Git.