Add FitNesseTestServerHostConsole project
The FitNesseTestServer can now be hosted in two different ways:

1) In a FitNesse test page, via the HttpServerFixture.  This method
can be used for self-contained FitNesse acceptance tests of
RestFixture.Net and for test pages demonstrating its functionality;

2) In the FitNesseTestServerHostConsole console application.  The
console application allows side-by-side comparisons of RestFixture
and RestFixture.Net tests when requesting the same resources from the
same server.  It will be used during RestFixture.Net development to
fine-tune the behaviour of RestFixture.Net so it is identical to
RestFixture.